requirements.txt
openai
langchain
pywin32
streamlit
python-dotenv
pandas
numpy
requests
streamlit
chromadb
tiktoken
pymupdf
